refactor: move content moderator to service folder
